% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airpollution}
\alias{airpollution}
\title{Airpollution}
\format{
An object of class \code{data.frame} with 128 rows and 5 columns.
}
\usage{
airpollution
}
\description{
Airpollution
}
\examples{
\dontrun{
 airpollution
}
}
\keyword{datasets}
