% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locust}
\alias{locust}
\title{Locust}
\format{
An object of class \code{data.frame} with 3864 rows and 5 columns.
}
\usage{
locust
}
\description{
Locust
}
\examples{
\dontrun{
 locust
}
}
\keyword{datasets}
