% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.qbld}
\alias{model.qbld}
\alias{qbld}
\title{QBLD sampler!}
\usage{
model.qbld(
  fixed_formula,
  data,
  id = "id",
  random_formula = ~-.,
  p = 0.25,
  nsim,
  b0,
  B0,
  c1,
  d1,
  method = "block",
  burn = 0,
  summarize = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{fixed_formula}{: a description of the model to be fitted of the form 
response~fixed effects predictors i.e Xi in the model.}

\item{data}{: data frame, NAs not allowed and should throw errors, factor variables are auto-converted, 
find airpollution.rda and locust.rda built into the package.}

\item{id}{: variable name in the dataset, that specifies individual profile. By default, id="id" and
data is expected to conatin an id variable.}

\item{random_formula}{: a description of the model to be fitted of the form 
response~random effects predictors i.e Si in the model. This defaults to Si being only an intercept.}

\item{p}{: quantile for the AL distribution on the error term, p=0.25 by default.}

\item{nsim}{: number of simultions to run the sampler.}

\item{b0, B0}{: Prior model parameters for Beta. See README for details.}

\item{c1, d1}{: Prior model parameters for Varphi2. See README for details.}

\item{method}{: Choose between the "block" vs "Unblock" sampler, regex is used to counter 
the effects of small/large cases.}

\item{burn}{: number between (0,1). Burn-in values are discarded and not used for summary calculations}

\item{summarize}{: Outputs a summary table (same as summary(output)), in addition also prints Model fit
AIC/BIC/Log-likelihood values (same as calling the 'mofit' function with appropriate inputs). False by default.}

\item{verbose}{: False by default. Spits out progress reports while the sampler is running.}
}
\value{
\itemize{
\item {\code{model.qbld}} {Returns `qbld` class object.}
}
}
\description{
Runs the QBLD sampler and outputs a qbld class object which consists of a list of markov chains
for Beta(the fixed effects estimate), Alpha(the random effects estimate), and Varphi2 (as per the model),
of which Beta and Varphi2 are of interest.
}
\details{
For a detailed information on the sampler, please check README.pdf 
data are contained in a data.frame. Each element of the data argument must be identifiable by
a name. The simplest situation occurs when all subjects are observed at the same time points. The
response variable represent the individual profiles of each subject, it is expected a variable in the
data.frame that identifies the correspondence of each component of the response variable to the
subject that it belongs, by default is named id variable. Hence NA values are not valid.

`qbld` object contains markov chains and sampler run information as attributes , and is compatible 
with S3 methods like summary,plot. make.qbld function can be used to convert a similar
type-object to `qbld` class. Check qbild/qbild2.pdf and the code to check for different additional options
in summary and plots.
}
\examples{
data = load("airpollution")
data2 = load("locust")
output <- (fixed_formula = wheeze~smoking+I(age^2)-1, data = data, id="id", 
            random_formula = ~-., p=0.25, 
           nsim=1000, method="block", burn=0, 
           summarize=FALSE, verbose=FALSE)
summary(output)
plot(output) 
output2 <- (fixed_formula = move~sex+I(time^2), data = data2, id="id", 
            random_formula = ~-., p=0.50, 
           nsim=1000, method="block", burn=0, 
           summarize=TRUE, verbose=FALSE)

}
\references{
Rahman and Vossmyer 2019, https://home.iitk.ac.in/~marshad/RahmanVossmeyer2019.pdf?
}
