\name{qbild-package}
\alias{qbild-package}
\alias{qbild}
\docType{package}
\title{
\packageTitle{qbild}
}
\description{
\packageDescription{qbild}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qbild}
\packageIndices{qbild}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{qbild}

Maintainer: \packageMaintainer{qbild}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
